pyproject.toml
[project]
name = "neo4j-database-agent"
version = "1.0.0"
description = "Neo4j Database Agent with interactive visualizations"
authors = [
    {name = "Agent Neo4j", email = "noreply@example.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk[eval]>=1.5.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.11.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.0.0",
    "networkx>=3.0.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "flask>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
eval = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "google-adk[evaluation]>=1.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["neo4j_database_agent*"]

[tool.setuptools.package-data]
neo4j_database_agent = ["*.py"]

[project.entry-points."google.adk.agents"]
neo4j_database_agent = "neo4j_database_agent.agent_factory:deployable_agent"

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']
